Refacto / Correction needed

+++ Umshini/Pilot/PartsType.cs
@@ -2,8 +2,8 @@
 {
     public enum PARTS_TYPE
     {
-        LEFT_WEAPON = WEAPON_SIDE.LEFT_WEAPON,
-        RIGHT_WEAPON = WEAPON_SIDE.RIGHT_WEAPON,
+        LEFT_WEAPON = (int)WEAPON_SIDE.LEFT_WEAPON,
+        RIGHT_WEAPON = (int)WEAPON_SIDE.RIGHT_WEAPON,
         FURNACE,
         LEGS
     }
